requirements.txt
Pillow==11.3.0
moviepy==1.0.3
instagrapi==2.2.1
openai==1.107.2
schedule==1.2.2
imageio==2.37.0
imageio-ffmpeg==0.6.0